[PRO-1249] Refactor MessageBusListenerActor out to common-message
